Create head component
