+++ packages/docs/src/components/layout.js
@@ -2,39 +2,21 @@
 import { jsx, Styled, Layout, Main, Box, Container } from 'theme-ui'
 import { useState, useRef } from 'react'
 import { Global } from '@emotion/core'
-import { Helmet } from 'react-helmet'
-import pkg from 'theme-ui/package.json'
 
 import SkipLink from './skip-link'
 import Header from './header'
 import Footer from './footer'
 import Sidebar from './sidebar'
 import Pagination from './pagination'
+import Head from './head'
 
 export default props => {
   const [menuOpen, setMenuOpen] = useState(false)
   const nav = useRef(null)
 
-  const title = [
-    props._frontmatter ? props._frontmatter.title : false,
-    'Theme UI'
-  ]
-    .filter(Boolean)
-    .join(' — ')
-
   return (
     <Styled.root>
-      <Helmet>
-        <title>{title}</title>
-        <meta name="description" content={pkg.description} />
-        <link rel="icon" type="image/png" href="/icon.png" />
-        <link rel="apple-touch-icon" type="image/png" href="/icon.png" />
-        <meta name="twitter:card" content="summary" />
-        <meta name="twitter:site" content="@jxnblk" />
-        <meta name="twitter:image" content="https://theme-ui.com/icon.png" />
-        <meta name="twitter:title" content={title} />
-        <meta name="twitter:description" content={pkg.description} />
-      </Helmet>
+      <Head {...props} />
       <Global
         styles={{
           '*': {
